Fix a small bug in #403 (#426)
* fi

* refactor

* fix
